# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NEXCART E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description?: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  imageUrl: string;
  images?: string[];
  brand: string;
  categoryId: number;
  categoryName: string;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  stockCount: number;
  tags?: string[];
  isFeatured: boolean;
  isDeal: boolean;
  createdAt: string;
}

export type ProductDetailSpecifications = {[key: string]: string};

export type ProductDetail = Product & ({
  specifications?: ProductDetailSpecifications;
  highlights?: string[];
  /** @nullable */
  warranty?: string | null;
  seller?: string;
});

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  icon?: string | null;
  productCount: number;
  /** @nullable */
  parentId?: number | null;
}

export interface CartItem {
  id: number;
  productId: number;
  product: Product;
  quantity: number;
  price: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  discount: number;
  deliveryFee: number;
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
  sessionId?: string;
}

export interface CartItemUpdate {
  quantity: number;
  sessionId?: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  sessionId: string;
  status: OrderStatus;
  items: CartItem[];
  subtotal: number;
  total: number;
  deliveryAddress: string;
  paymentMethod: string;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
}

export type OrderInputItemsItem = {
  productId: number;
  quantity: number;
  price: number;
};

export interface OrderInput {
  sessionId: string;
  deliveryAddress: string;
  paymentMethod: string;
  items: OrderInputItemsItem[];
}

export interface Review {
  id: number;
  productId: number;
  rating: number;
  title: string;
  body: string;
  author: string;
  verifiedPurchase: boolean;
  helpful: number;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  title: string;
  body: string;
  author: string;
}

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  bgColor: string;
  ctaText: string;
}

export interface StoreSummary {
  totalProducts: number;
  totalCategories: number;
  totalDeals: number;
  topBrands: string[];
  avgRating: number;
}

export interface ChatMessageInput {
  message: string;
  sessionId: string;
  /** @nullable */
  context?: string | null;
}

export interface ChatResponse {
  reply: string;
  suggestions: string[];
  productIds: number[];
}

export interface WishlistAction {
  sessionId: string;
}

export interface SessionRef {
  sessionId?: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
minPrice?: number;
maxPrice?: number;
brand?: string;
rating?: number;
sort?: ListProductsSort;
page?: number;
limit?: number;
inStock?: boolean;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
  newest: 'newest',
  popular: 'popular',
} as const;

export type GetSearchSuggestionsParams = {
q: string;
};

export type GetCartParams = {
sessionId?: string;
};

export type ClearCartParams = {
sessionId?: string;
};

export type GetWishlistParams = {
sessionId?: string;
};

export type ListOrdersParams = {
sessionId?: string;
};