# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXCART E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{slug}/products:
    get:
      operationId: getCategoryProducts
      tags: [categories]
      summary: Get products by category slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Products in category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get the current cart
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRef"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/clear:
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear the entire cart
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for a session
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountPercent:
          type: ["integer", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        brand:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        stockCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isDeal:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - price
        - imageUrl
        - brand
        - categoryId
        - categoryName
        - rating
        - reviewCount
        - inStock
        - stockCount
        - isFeatured
        - isDeal
        - createdAt
...
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, totalPages]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
      required: [id, productId, product, quantity, price]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required: [items, subtotal, discount, deliveryFee, total, itemCount]
    CartItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        sessionId:
          type: string
      required: [productId, quantity]
    CartItemUpdate:
      type: object
      properties:
        quantity:
          type: integer
        sessionId:
          type: string
      required: [quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        total:
          type: number
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
        estimatedDelivery:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - status
        - items
        - subtotal
        - total
        - deliveryAddress
        - paymentMethod
        - createdAt
    OrderInput:
      type: object
      properties:
        sessionId:
          type: string
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              price:
                type: number
            required: [productId, quantity, price]
      required: [sessionId, deliveryAddress, paymentMethod, items]
...
    SessionRef:
      type: object
      properties:
        sessionId:
          type: string