# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NEXCART E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  sessionId: string;
  status: OrderStatus;
  items: CartItem[];
  subtotal: number;
  total: number;
  deliveryAddress: string;
  paymentMethod: string;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
}