# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NEXCART E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';

export interface OrderInput {
  sessionId: string;
  deliveryAddress: string;
  paymentMethod: string;
  items: OrderInputItemsItem[];
}