# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NEXCART E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  description?: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  imageUrl: string;
  images?: string[];
  brand: string;
  categoryId: number;
  categoryName: string;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  stockCount: number;
  tags?: string[];
  isFeatured: boolean;
  isDeal: boolean;
  createdAt: string;
}